requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dateutil==2.8.2

# AI/ML
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1

# Data processing
pandas==2.1.3

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Date utilities
python-dateutil==2.8.2

# Pydantic for settings
pydantic==2.5.1
pydantic-settings==2.1.0
email-validator==2.1.0






